Refactoring Decanter::Core
This is rather a large refactor, hopefully simplifying some code paths.

I've added a #handle method that can be used to create generic handlers,
if for some reason :input, :has_many, :has_one don't fit the bill.

I've also added an `:as` option, allowing parameters with names other
than their associated method to be parsed.  (Need to add this to the
README).

Finally, I've added tests that reflect the examples in the README.